fiz o header na master sem perceber mas está funcionando normalmente
