+++ pokedex/src/pages/DetalhesPokemon.js
@@ -1,9 +1,11 @@
 import React from 'react'
 import { Box, Text } from "@chakra-ui/react"
+import Header from './Header'
 
 const DetalhesPokemons = () => {
   return (
       <Box>
+        <Header/>
         <Text>Detalhes dos Pokemons</Text>
       </Box>
   );
