+++ pokedex/src/pages/Home.js
@@ -1,9 +1,16 @@
+import { Heading, Box, Text } from '@chakra-ui/layout'
 import React from 'react'
-import { Box, Text } from "@chakra-ui/react"
+import { useHistory } from "react-router-dom"
+import Header from './Header';
+
 
 const Home = () => {
+
+  const history = useHistory();
+
   return (
       <Box>
+        <Header/>
         <Text>Home</Text>
       </Box>
   );
