+++ pokedex/src/pages/Pokedex.js
@@ -1,9 +1,11 @@
 import React from 'react'
 import { Box, Text } from "@chakra-ui/react"
+import Header from './Header'
 
 const Pokedex = () => {
   return (
       <Box>
+        <Header/>
         <Text>Pokedex</Text>
       </Box>
   );
